requirements.txt
pymupdf4llm
marker-pdf
ollama